BlogServiceTests

+++ app/Services/BlogService.php
@@ -28,11 +28,13 @@ class BlogService {
         $blog->title = $title;
         $blog->content = $content;
         if ($image){
-            $this->imageService->removeExistingImage($blog->image);
+            if ($blog->image)
+                $this->imageService->removeExistingImage($blog->image);
             $filename = $this->imageService->getFileNameFromRequestAndSaveIt($image);
             $blog->image = $filename;
         }
         $blog->countries = $countries;
         $blog->update();
+        return $blog;
     }
 }
